Update patch set 2
Patch Set 2:

Ned, Ben and I sat and discussed the architecture of the solution offline and realized that, as it stands today, the workaround I proposed is probably the simplest way for metric authors to derive the benefit of this CL. So in other words, the leak detection metric would say something like:

const leaks = findLeaks();
addHistogramsForLeaks(leaks, histograms);
if (leaks.length !== 0) {
  throw new Error('There were leaks');
}

As of right now, this will already cause an MRE Failure to be generated, which will be turned into a FailureValue in Telemetry, which will in turn cause the story to be marked as failed.

Does that work for you? If so, let's just suggest this to Yuzu and I'll abandon this CL until we have additional use cases for this approach. If you still have concerns, we can discuss them on Monday.

Patch-set: 2
Reviewer: Gerrit User 1116849 <1116849@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>