Translate HistogramGrouping to python.
Currently, Histograms can only be grouped and merged in Javascript.
The Histogram Pipeline passes results through vinn in order to merge
Histograms across storyset repeats and stories. The merge_histograms
script would be more efficient if it did not need to pass through vinn.
Also, if users are to be able to process Histograms in colab, they will
need a pure-python implementation of merging.

This CL starts to implement grouping and merging Histograms in Python.
This is very close to a line-by-line translation from the JS
implementation of HistogramGrouping (currently used in results.html)
to Python.

Followups:
translate HistogramSet.GroupHistogramsRecursively(),
translate HistogramSet.DeduplicateDiagnostics(),
translate merge_histograms.

Bug: catapult:#3806

Change-Id: Iedbacc6a84faf2ce7ca18c024dac8cdb0f23c654