Update Python Histogram.Clone() to reflect JS implementation.
Currently, Histograms can only be merged in Javascript.
The Histogram Pipeline passes results through vinn in order to merge
Histograms across storyset repeats and stories. The merge_histograms
script would be more efficient if it did not need to pass through vinn.
Also, if users are to be able to process Histograms in colab, they will
need a pure-python implementation of merging. merge_histograms relies
on the semantics of the Javascript Histogram.clone(), which is more
modern than the old CloneEmpty() in the Python Histogram.

This CL updates the Python Histogram.Clone() to reflect the modern
semantics of the Javascript Histogram.clone() as required by the
merge_histograms algorithm.

Bug: catapult:#3806

Change-Id: I93a60dc92f49fa8ca6e96b234a0784f5c0fba8b8