Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > I am confused, how do you run performance tests on APKs on desktop environment?
> > 
> > I run them on Android, but catapult will try to match the APK to all known environments, including desktop. When the desktop matcher runs, it throws this exception, which terminates the entire matching process.
> 
> Oh I see. Is that still the case if you run with --device=android?

If I explicitly pass a device serial number with --device it does seem to work. That works for me, but it may be worth changing the wording of the exception so that it is clear that that is what one needs to do.

Patch-set: 4
Reviewer: Gerrit User 1125980 <1125980@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0