[pinpoint] Change "differences" section wording.
Originally, we had thought of "Differences found after commits" as being the
section header, so it made sense fo the section contents to be, "No differences
found." But if you don't realize at first that it's the section header, the
apparent contradiction is confusing. So instead, make the header change to "No
differences found".

Bug: catapult:#4037
Change-Id: I6389eba47c766e27b36ae9d83eebfa63aa0ac058