Don't log BattOr errors from generating the serial map
One very common failure in generating the serial map is that there's a
phone without a connected BattOr, which happens at the start of all
benchmarks where BattOr is enabled but on a configuration without
BattOrs present. Logging the exception with the current traceback gives
a somewhat misleading idea that this exception is unexpected.

Since logging this exception, we've added a second, gentler
message: "BattOr power monitor is [not ]connected."

Change-Id: I3777c723b90641eb5256e527e1928b6a33b5425a