Update patch set 38
Patch Set 38:

Sorry missed the discussion since I was out, what were the concerns that having an existing file checked in was preferred?

We had a similar question pop up for Pinpoint when we were integrating results2 for perf try jobs. The concern at that time with checking in a vulcanized version was largely around who was supposed to keep it up to date, and opted to have Pinpoint vulcanize as part of the deployment step.

Patch-set: 38
Reviewer: Gerrit User 1116018 <1116018@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>