Make each soundwave dashboard API request use a new connection
This is a step towards making requests to the dashboard API in parallel.
App engine credentials are thread-safe, but connections are not. Because
connections are pretty lightweight to create, we can stash the
credentials instead of a connection and just use that credentials object
to create a new connection each time we make a request to the dashboard.
This makes the move to multithreading much easier.

Change-Id: Ic37bf5c1e1058d664a2bb33ec5f02791aa40c3be