Make pinch actions take speed in viewport coords
I was a little over-zealous in
https://crrev.com/dbab215523c8d77a7b54fcf3776357b079c36899. I made the
speed parameter in PinchAction be in CSS pixels. This makes sense for
scrolls and other gestures but for pinch it means that the speed of the
zoom depends on how far the page is currently zoomed in. This is never
what you want for pinch-zoom.

I've updated pinch.js to keep the speed invariant across page scale
factors. I've also added a test for this and other pinch zoom aspects
since we currently have poor test coverage. In fact, this exposed bugs
in our MotionEventSynthesizer for multi-touch. The second point DOWN
and UP events did not change the pointer action index (as returned by
getActionIndex()) which meant the ACTION_POINTER_DOWN event intended
for the second pointer was actually setting the action on the first.
This patch includes fixes for these issues as well.

Bug: chromium:610021
Change-Id: I7008cb6c90aa4b9a610abd2c322c88281aa97f87