Reland "Test that gestures scroll the correct distance"
This patch was originally landed in
7e4796650c837c13363eceb011d0c8bd8fef72dc
and then reverted in d0e5c9c00f4e0d4e654efd9eb9f3d8fa025a8a4e due to
failing on the Mac bot when rolled into the Chromium tree.

The issue in the original patch was that the tests were sending touch
events to a Mac browser which causes a crash. This CL relands that patch
by skipping touch tests on Mac.

Bug: chromium:778587
Change-Id: Ibb08baae2b70c42e092fc877f252c7073b0f8493