Dashboard - Cleanup unused masters and bots
We already have a cron job /deprecate_tests which removes unused
TestMetadata and Rows, but leaves the Master and Bot entities lingering
in datastore forever. This does cleanup for those.

For masters, if there is no bot with the master_key as an ancestor,
then we delete it.

For bots, if there are no TestMetadata entities that reference that bot
and that bot's master, we can delete the bot.


Bug: catapult:#4054
Change-Id: I4dec1fc456248fac03bf3ca486b053acbf5a299a