Dashboard - Defer gathering test paths in /edit_sheriffs.
Still getting timeouts on this endpoint. Longer term may need to
rethink sheriffs and how they're stored but for now punting this to
taskqueues to get the 10 min limit.

Also needed to break up querying for the specific test paths affected
since it was possible to hit the soft-memory limit doing this all at
once. This should run fine now, but this can potentially result in
more work since the old version could set(all_test_paths) and this
version can't.

Bug: catapult:#4027
Change-Id: Ie1a024ee66502a0c1ac73a2a1c1c34fa3d14cf82