[Telemetry] Remove PortKeeper
Previously (in AndroidBrowserBackend) the PortKeeper was used to
hold a port, reserved during `__init__`, to be later "released"
during `Start` and then immediately put into use by "adb foward".

Instead we now can call GetUnreservedAvailableLocalPort, which
both "holds" and "releases" a port to be immediatelly used by
"adb forward" during `Start`.

No approach is safer than the other: both suffer form a small
race as somebody else might get to use the port after releasing
and before forwarding starts on it.

The latter aproach, however, is cleaner. And will help to get
closer to a model where the port selection can be abstracted
away and delegated to a new "Forwarder" class.

Since AndroidBrowserBackend was the only user of PortKeeper,
we remove the PortKeeper entirely.

Bug: catapult:#1977
Change-Id: I484976c6034307d29363fde14d307dc7df083eed