Update patch set 4
Patch Set 4:

So, not an owner here, Annie or Ned can weigh in. If this is for your dashboard ui prototype, why not just upload the components for now and file a bug to discuss adding a hook or something to get the bower dependencies since how to do that is up in the air.

Also, do you have a tracker issue or anything for the prototype? Feel like this should be referring to that, rather than github:3058 which doesn't seem related.

Patch-set: 4
Reviewer: Gerrit User 1116018 <1116018@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>