Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> So, not an owner here, Annie or Ned can weigh in. If this is for your dashboard ui prototype, why not just upload the components for now and file a bug to discuss adding a hook or something to get the bower dependencies since how to do that is up in the air.

Also not an owner here, and don't feel too strongly about any of this.

I think a point to address even before we talk about this would be the point that we should probably use the Polymer CLI https://www.polymer-project.org/2.0/docs/tools/polymer-cli instead of bower.

I think maybe uploading the components is best for now, because the reality is that this approach will require anyone who wants to hack on Polymer 2 stuff in Catapult to have bower, which is not necessarily something we want to assume. I think getting to a point where we can use a Catapult-provided version of polymer-cli (via node_runner) to bootstrap Polymer dependencies is a laudable goal that we should backlog.

> 
> Also, do you have a tracker issue or anything for the prototype? Feel like this should be referring to that, rather than github:3058 which doesn't seem related.

Patch-set: 4
Reviewer: Gerrit User 1116849 <1116849@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>