Re-reland "Test that gestures scroll the correct distance"
The original patch in 7e4796650c837c13363eceb011d0c8bd8fef72dc was once
again reverted in 86c17b98f5364a8a3f903df3d946ea7afe4d729b. This time
the newly added test broke on Windows and Linux during the catapult
roll.

This test failure actually uncovered an issue in animated wheel
scrolling, the fix for which is in https://crrev.com/c/788134.
Additionally, I discovered that animated scrolls can lose some delta
when the visual viewport reaches the layout viewport extent. At that
point the layout viewport starts animating, however, the delta that
contributed to the visual viewport's scroll animation curve is lost.

This CL changes the testWheelScrollDistanceWhileZoomed test to first
scroll the visual viewport to the end so that the entire delta is
applied to the layout viewport.

Bug: chromium:778587
Change-Id: I50520fd1c43e607d1e79ce22b6a0a8ba9fee154e