Pinpoint - Always run ReadValue quest
Previously, if no chart was specified the ReadValue quest was skipped.
Problem is that try jobs don't necessary need to specify a metric, but
you still want it to read out all the values (and other data like trace
urls).

This forces it to run, which in turn reads the resulting histogram
output from the test, and parses the trace urls for display.

Demo: https://pinpoint-dot-chromeperf.appspot.com/job/14ea0ddff80000

Bug: catapult:#4048
Change-Id: I8140ea18034a89c6ed463f9e1643fca2b7ba9609