Update patch set 38
Patch Set 38:

> Patch Set 38:
> 
> Sorry missed the discussion since I was out, what were the concerns that having an existing file checked in was preferred?

Either we should support "having an existing file checked in", or the vulcanization. Only one of them should be good because any of the two ways will remove the need for other.

Last time we discuss, there is the use case of using this tool for google cloud, which vulcanization is not an option there, so I think we should only support the former.

To be clear, I am not an owner of this code base, so leave the final decision to you folks. It's just me advocating for "less is more" as usual :-)
> 
> We had a similar question pop up for Pinpoint when we were integrating results2 for perf try jobs. The concern at that time with checking in a vulcanized version was largely around who was supposed to keep it up to date, and opted to have Pinpoint vulcanize as part of the deployment step.

Patch-set: 38
Reviewer: Gerrit User 1128435 <1128435@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>