Dashboard - Resave Tests when sheriff differs.
This adds an extra step to /deprecate_tests which checks the existing
sheriff vs the ideal one, and resaves the test if they differ.

Due to timeouts in /edit_sheriffs, it's been easy for the patterns
and tests to get out of sync. Resaving all tests, and then spreading out
the work in /edit_sheriffs should fix things.

Bug: catapult:#4026
Change-Id: I568ff46a975dd8f20e0f25e62fdbe549f985081f