Fix another bug in the simplified dropdown.
When a checkbox is focused and the user presses spacebar, Chrome dispatches a
keydown, event, a keypress event, a change event, and a click event. The
checkbox created by tr.ui.b.createCheckBox() only listens for the change event,
so the click event bubbles up to the dropdown, which Polymer re-dispatches as a
tap event. Since it was caused by a keypress, its coordinates are (0,0) which is
outside of the dropdown so it closes. Before my recent change to simplify
dropdown, this bug was prevented by this line:
if (e.srcElement !== this.$.dialog) return;

This CL restores that line of code, and updates the test.

Bug: catapult:#789879
Change-Id: I0346d11d4ac7cffc47f19c6e4a0f32bf6d1ca7bc