Update patch set 38
Patch Set 38: Code-Review+1

Simon and Ben will discuss this so that Simon can form an opinion, but my two cents on the vulcanization vs. template approach is that having both is preferable to having only the template approach. In particular, if we take the approach of checking a template into Catapult that we update in PRESUBMIT or something like that, then that effectively means that _any_ change in tracing/ will require users to update the template. I'm more comfortable with the idea that users who want to use the vulcanized template option will develop their own deployment strategy.

Patch-set: 38
Reviewer: Gerrit User 1116849 <1116849@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1