Fix rendering of full height UI features
Canvas pagination breaks full height UI features drawn by the model track
(since 78b00b99c "Constrain the max size of the canvas").

Once the user scrolls down far enough, the rendering canvas is moved down
the page. This canvas movement ends up reflected in the global
translation applied to all rendering so that scroll dependent content
ends up in the same place.

However, not all rendering is intended to be scroll dependent. In
particular full height UI such as interest range shading need to look the
same regardless of scroll offset. These features intend to paint from the
top of the canvas to its bottom, but currently end up clipped because
they are not aware that an upwards translation can be applied even to
rendering within the top level "model" track.

Any trace taller than 3 times the size of the viewport will exhibit this
issue, particularly full system traces collected by Chrome OS & Android.

This switches all uses of ctx.canvas.height to use the track height
instead, so that coordinates used while rendering is consistently
independent of the vertical positioning of the canvas. When pagination
occurs, the model is larger than the canvas, and the draw calls will
be clipped to the right size.

Bug: catapult:#3266
Change-Id: I7f753a1be62bb615601746ab670579bb7c8d6604