Add CompositorThread to GPU process thread names.
When running with --enable-viz code is shuffled around between
processes. Some trace events for smoothness benchmark are output in the
GPU process on the 'CompositorThread' with --enable-viz. Telemetry
doesn't look for 'CompositorThread' and misses all the relevant trace
events as a result.

Update TraceEventTimelineImporter to look for 'CompositorThread' when
it's trying to identify the GPU process.

Bug: chromium:778749
Test: Verified telemetry works with --enable-viz.
Change-Id: Iea2bcc1cdc412c1ea39e0d3c0f008977c3725563