Add about_tracing.html to quick links
The dev server is useful way to test bug fixes to chrome://tracing but
the URL is not very discoverable currently. Add a link to the "quick
links" section.

There's a version at "Trace File Viewer" but it is awkward to use; it has
a dropdown of canned traces instead of a "Load" button so you have to
copy all your traces into the dev server directory before loading them.

Change-Id: Idffe55612719975d18bb53c816c1d3e023648210