Pinpoint - Grab histogram summary metric equivalent.
When no story is supplied in the chartjson path, we look for a summary
metric but on the histograms side we currently try to match against
histograms with no story. Instead, if no story is supplied, we skip
looking at histogram's story diagnostic, which gives us all the data
for a given histogram/tir_label, which is what summary would have done.

Bug: catapult:#4060
Change-Id: I4ae27498095cf3082bee593912c10a58c5eba340