tracing: Fix length calculation in base64 module
Base64.DecodeToTypedArray() only returns the correct length if the
decoded length is one less than a multiple of three, such as in the test
string "hello". The other two cases either overestimate or underestimate
the length by 1. The empty string results in a length of -1.

One out of every three traces should have lost its last byte in the
trace2html importer because of this. Hopefully it was whitespace.

Fix it by making all lengths exact.

BUG=chromium:786091

Change-Id: I9a88a4e2c7ed78ece782b91d1c90cb66817d259c