Remove platform_backend's DidStartBrowser, DidCreateBrowser and WillCloseBrowser hooks
The only part of these hooks that create actual effect on perf test is
the calls to SetFullPerformanceModeEnabled in DidCreateBrowser & unset it in
WillCloseBrowser. As we remove those hooks, those calls are moved to
shared_page_state for now (ideally they should be invoked in story_runner so
other shared_state can reuse this, but
the code there is already way to complexed, so I just leave it there for now).

The life cycle of SetFullPerformanceModeEnabled now is:

SetFullPerformanceModeEnabled(...) is called before any test is run (before
browser is started) in SharedPageState.__init__

SetFullPerformanceModeEnabled(False) is called after all the test has run
in SharedPageState.TearDownState().

Compared with before, we run SetFullPerformanceModeEnabled less often & this is
a good optimization as there is no need to set performance mode many times.

Notes that this also changes relative order of the ClearCaches &
SetFullPerformanceModeEnabled as following:

Before:
ClearCaches() (in desktop_browser_finder.Create)
SetFullPerformanceModeEnabled(...) (in browser.__init__)

After:

SetFullPerformanceModeEnabled(...) (in shared_state.__init__)
ClearCaches() (in desktop_browser_finder.Create, in shared_state.WillRunStory)

Hence with this CL, the relative order of these two events would be back to what they were before
the regression in crbug/792357

Bug: chromium:792860
Bug: chromium:792357
Change-Id: Ibeba9ca0af55921514d4f3fecf692c3a740b1aac