Adds new interactivity metrics to loading metrics
This CL replaces the old TTI definition in loading metric with the new
TTI definition, and also adds the Time to First CPU Idle metric. Since
there is a change in the structure of diagnostic info, it updates all
the code locations that use |collectLoadingMetricsForRenderer|.

Since the definition of TTI is used by expected queuing time and power
metrics, and also to define Successful Load user expectation, this CL
might cause significant changes in values on the perf dashboard.

Time to Interactive and Time to First CPU Idle are the recently renamed
names of Time to Consistently Interactive and Time to First Interactive
respectively. See https://goo.gl/vdfeegmW for rationale for these metric
definitions.

Bug: catapult:#3635
Change-Id: Ia12ad0f6d09303a9e6ad0af124df3a1aa0a72c00