[Telemetry] Fix browser_backend.pid for Android
The old code used to expect finding a single process matching the name
of the browser app.

This, however, is not always reliable in cases where the browser
forks its main process and, for a short period of time, two processes
may be found with the same name.

Instead we now look for the single process with the matching name
which is also a direct descendant of the Android zygote.

Bug: chromium:785446
Change-Id: I0724c382118a4ef1f98d970bdf0f44abf1494a58