Update patch set 1
Patch Set 1:

Actually in record mode, people are supposed to ignore the test expectation. The reason is they are going to re-record the page (rebuild it) anyway. So if the test is broken, they can fix it right away or just remove it completely.

Patch-set: 1
Reviewer: Gerrit User 1128435 <1128435@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>