Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Actually in record mode, people are supposed to ignore the test expectation. The reason is they are going to re-record the page (rebuild it) anyway. So if the test is broken, they can fix it right away or just remove it completely.

That is not how it has ever worked though. Even with annotation disabling it required --also-run-disabled-tests to record disabled tests. We can certainly make it ignore it... but if we do want that behavior we have several options in how to achieve it.

We could just not have record_wpr pass the expectation file. That would just mean not landing this cl:
https://chromium-review.googlesource.com/c/chromium/src/+/822983

Patch-set: 1
Reviewer: Gerrit User 1132400 <1132400@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>