Update patch set 1
Patch Set 1:

> Is it possible to change record_wpr to run all the tests? The previous behavior probably wasn't thought through well.

How does record_wpr react when chrome crashes in the middle of the recording run? Does it just not record that story? Or does it cause the entire set of recordings to be aborted? If we cannot 100% promise the first behavior, we should probably keep it as is. It has a sensible override to let you record disabled stories. The other option would be to add the opposite flag to record_wpr so you can rerecord around hard failures.

Patch-set: 1
Reviewer: Gerrit User 1132400 <1132400@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>