Update patch set 1
Patch Set 1:

> Breakage like Chrome crash or story error should not abort the whole recording (or benchmark run).

My worry is that even though we intend that to be the behavior, it doesn't mean that its always the behavior we get. I know there was a problem awhile back with BattOr stories that when they failed, they failed hard even though they weren't expected to. Is that bug fixed?

Either way, if later on someone needs to implement this for some reason, it wont be hard. I'll abandon this CL, and start on another just removing all expectation use in record_wpr.

Patch-set: 1
Reviewer: Gerrit User 1132400 <1132400@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>