[pinpoint] Allow RunTest to use "configuration" or "dimensions".
For the "perf try job dialog", users will want the ability to specify bots by
configuration instead of dimensions. So we should do the conversion on the
backend side.

The unit tests are a little bit messy right now. In a future CL I will likely
move the quest generator functions into the individual quests, which will clean
up the testing here too.

Bug: catapult:#4111
Change-Id: Iae11e9a889df94454cb9b13e56910f7ed4bb8a79