Dashboard - Some optimizations to speed up /add_histograms.
Estimated we'd take about 110s or so to process an upload from
media.desktop. A lot of this was in task creation, specifically in
creating tasks. Using params gets urlencoded, so switching to payload
to just dump the values into body. Also, we estimate the size of the
task ourselves by keeping track of the json'd data.

Overall brings endpoint down to 30-35s for the media.desktop case. The
majority of the time now is in various histogram related functions
(importing, replacing diagnostics, etc).

Bug: catapult:#4118
Change-Id: I632aaeebb8472064d61b07969f375c0b69cd0470