Update patch set 2
Patch Set 2:

Ned, PTAL. This is still WIP, but wanted to get your thoughts on this approach before going forward.

Currently _WaitForBrowserToComeUp is a bit of a misnomer, because more than waiting for the browser, is the key piece responsible for binding a devtools client running on the target platform to the browser object just started.

My plan is to make this explicit by splitting this into two parts:

- _GetDevToolsClientConfig - platform specific, figures out which are the local/remote ports for the devtools client, and the browser_target needed. On Android, crucially, this pretty much only depends on knowledge from the browser_backend settings.

- BindDevToolsClient - same for all platforms, after getting the config from the previous method, waits-for and creates the devtools client object.

So this would later allow to write on PossibleAndroidBrowser:

def FindBrowserInstance(self):
  browser_backend = # create the browser backend, same as today Create
  browser_backend.BindDevToolsClient()
  return browser.Browser(browser_backend, ...)

Again, this works on Android because BindDevToolsClient does not need any information about the running process (other than the package name, and for webview the PID which we can also find by package name).

[Although not a goal, something like FindBrowserInstance could even work on Desktop. It's a bit tricky because the current implementation assumes we've got hold of the _proc started by subprocess.Popen, but may be doable if needed at some point in the future.]

Let me know if this sounds reasonable. Still missing are breaking apart the _WaitForBrowserToComeUp in the cros browser, and a small refactor in DevToolsClientBackend.

Patch-set: 2
Reviewer: Gerrit User 1128435 <1128435@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1123010 <1123010@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false