tracing: Add support for system tracing via chrome://inspect?tracing
Tracing via remote debugging (chrome://inspect) cannot support system
tracing or any other secondary tracing agents supported by chrome
tracing. The option is shown in the UI but the collected system trace
events will not be visible in the returned data.

This is due to the limitations of the devtools protocol's "ReportEvents"
transfer mode, which streams individual chrome trace events. System
tracing events are embedded into the trace.json format as a blob of text,
not the chrome event format.

The "ReturnAsStream" transport delivers the full trace.json including any
embedded system trace events, so change the inspector tracing controller
to prefer this mode. This will ensure feature parity between features
supported by chrome://tracing and chrome://inspect?tracing going forward.

Also opt into compression if it is available, as compressed traces are
typically 1/10th the size of uncompressed ones.

This is backwards compatible with devtools endpoints that do not support
gzip, as well as old endpoints that do not support the "ReturnAsStream"
transfer mode.

BUG=chromium:786091

Change-Id: I8fa200b371d5237f62140c88160cd250a7c84fed