Update patch set 2
Patch Set 2:

It seems odd that cavalcantii@ is unable to add this patch to the CQ. Does this project have a different permissions list to the chromium/src one? 

In any case, I've addressed the nit, but am unable to submit the patch to the CQ myself. Could someone do it for me, please?

Patch-set: 2
Reviewer: Gerrit User 1190714 <1190714@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>