Increase a couple of Telemetry's internal timeouts to 60s.
The first WebSocket connection to the browser, and fetching of
SystemInfo, depends on a lot of work being done in the target browser,
and in Linux Debug builds on the bots, these timeouts (of 10s in some
cases) were being hit.

Explicitly indicate which are the likely first calls to the newly-
started browser, and increase their timeouts.

It seemed difficult to figure out the browser configuration
(Release/Debug) at this level, so do this in all cases.

Make GetSystemInfo's timeout an optional argument, stop specifying it
at the call site, and use the increased timeout on all platforms.

BUG=chromium:796437

Change-Id: I34426a5a30dbc2231fc740476cbdcf343a2d1ffb