Make pinch actions take speed in viewport coords
I was a little over-zealous in
https://crrev.com/dbab215523c8d77a7b54fcf3776357b079c36899. I made the
speed parameter in PinchAction be in CSS pixels. This makes sense for
scrolls and other gestures but for pinch it means that the speed of the
zoom depends on how far the page is currently zoomed in. This is never
what you want for pinch-zoom.

I've updated pinch.js to keep the speed invariant across page scale
factors. I've also added a test for this and other pinch zoom aspects
since we currently have poor test coverage.

Note: This is re-landing b83a27983a06bea1f8a94ad8b3b56c2ef6279d1b

Bug: chromium:610021
Change-Id: I9444be9929b9ce16d9b4d65eb2be405bb99d3a24