[Tracing] Fix flaky test caused by flaky gsutil.
This fixes the native heap symbolizer test in catapult
by fixing the oauth2client used by gsutil used by the test.

The default timeout is 0, and when timeout is 0 it does not
do retries for busy files. Any thread that uses gsutil locks
against the same file, so we get flaky failures because if the
lock is taken it will never do a retry.

Bug: catapult:#4134
Bug: chromium:796137
Change-Id: I903f4fabc3f228326fdf8982896001d6d817cb46