Remove mentions of unused trace event
The trace event "Startup.BrowserProcessCreation" indicates the process
creation time (on Linux by reading /proc/self/stat), but does not seem
to be used anywhere in Telemetry.

For future benchmarking it seems more reliable to use TimeTicks::Now()
as it is based on monotonic clock on all systems. The only disadvantage
of it is that it does not cover the time spent before reaching the
main entry point (aka "main()" on non-Android).

Bug: chromium:795836
Change-Id: I7f4bf78edb69d12449fbbf00b7e31c607a83ff63