Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 5:
> > > 
> > > lgtm, but I don't have permission to set Code-Review.
> > 
> > It seems that no-one of the reviewers has "Code-Review" votable for this CL!
> 
> Achuith is owner for this code, but he is OOO.
> Ethan also could approve this code, but he is not responding.

I am not qualified to approve this code. Please note that in the Telemetry OWNERS file https://github.com/catapult-project/catapult/blob/master/telemetry/OWNERS#L7 it states that I review changes to do with the results and value systems.

If this is so urgent that it can't wait until after the holiday break, then you should TBR Achuith, because this is a cros-specific fix. When the "full" Telemetry owners are back in office, you should work with them to establish more up-to-date per-file ownership of these files.

Patch-set: 5
Reviewer: Gerrit User 1116849 <1116849@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>